requirements.txt
langgraph>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.25.0
uvicorn[standard]>=0.20.0
langchain_groq>=0.1.0